Revert "Revert "Story cards #5""
