+++ app/components/AllStories.jsx
@@ -1,13 +1,12 @@
 import React, { Component } from 'react'
 import SingleStory from './SingleStory'
-import Navbar from './Navbar'
 import firebase from 'app/fire'
 import 'firebase/database'
 import {fetchAllStories} from '../reducers'
 import {connect} from 'react-redux'
 import _ from 'lodash'
 
-class AllStories extends Component{
+class AllStories extends Component {
   componentWillMount() {
     this.props.fetchAllStories()
   }
@@ -15,8 +14,7 @@ class AllStories extends Component{
   render() {
     const {allStories} = this.props
     return (
-      <div>
-      <Navbar />
+      <div className="row">
         <div className="col-sm-4 col-md-4 col-lg-4" >
           {
             !_.isEmpty(allStories) &&
