+++ app/components/NewBranch.jsx
@@ -1,7 +1,7 @@
-import React from 'react';
-import Dialog from 'material-ui/Dialog';
-import FlatButton from 'material-ui/FlatButton';
-import RaisedButton from 'material-ui/RaisedButton';
+import React from 'react'
+import Dialog from 'material-ui/Dialog'
+import FlatButton from 'material-ui/FlatButton'
+import RaisedButton from 'material-ui/RaisedButton'
 
 export default class DialogExampleSimple extends React.Component {
   state = {
@@ -9,11 +9,11 @@ export default class DialogExampleSimple extends React.Component {
   };
 
   handleOpen = () => {
-    this.setState({open: true});
+    this.setState({open: true})
   };
 
   handleClose = () => {
-    this.setState({open: false});
+    this.setState({open: false})
   };
 
   render() {
@@ -29,7 +29,7 @@ export default class DialogExampleSimple extends React.Component {
         keyboardFocused={true}
         onClick={this.handleClose}
       />,
-    ];
+    ]
 
     return (
       <div>
@@ -44,6 +44,6 @@ export default class DialogExampleSimple extends React.Component {
           The actions in this window were passed in as an array of React objects.
         </Dialog>
       </div>
-    );
+    )
   }
 }
