+++ app/routes.js
@@ -4,6 +4,9 @@ import {Route, Switch} from 'react-router-dom'
 // authentication
 import NotFound from './components/NotFound'
 
+import Navbar from './Navbar'
+import Footer from './Footer'
+
 import LandingPage from './components/LandingPage'
 import WriteSpace from './components/WriteSpace'
 import AllStories from './components/AllStories'
@@ -13,15 +16,21 @@ import UserProfile from './components/UserProfile'
 class Routes extends Component {
   render() {
     return (
-      <Switch>
-        <Route exact path="/" component={LandingPage} />
-        <Route path="/home" component={LandingPage} />
-        <Route path="/write" component={WriteSpace} />
-        <Route exact path="/read" component={AllStories} />
-        <Route path="/read/:id" component={SingleCommunity} />
-        <Route path="/userProfile" component={UserProfile} />
-        <Route path='*' component={NotFound} />
-      </Switch>
+      <div>
+        <Navbar />
+        <div className="pageContent">
+          <Switch>
+            <Route exact path="/" component={LandingPage} />
+            <Route path="/home" component={LandingPage} />
+            <Route path="/write" component={WriteSpace} />
+            <Route exact path="/read" component={AllStories} />
+            <Route path="/read/:id" component={SingleCommunity} />
+            <Route path="/userProfile" component={UserProfile} />
+            <Route path='*' component={NotFound} />
+          </Switch>
+        </div>
+        <Footer />
+      </div>
     )
   }
 }
