+++ app/stylesheets/style.scss
@@ -1,0 +1,94 @@
+@import '/css/bootstrap.min.css';
+
+@media (max-width: 768px) {
+  .btn, .btn-group {
+      width:100%;
+  }
+}
+
+html, body {
+  font-family: "Roboto", sans-serif;
+  font-weight: 300;
+  margin: 0;
+  padding: 0;
+  background-color: #FAFBF6;
+}
+// below html & body needed for sticky footer
+html {
+  position: relative;
+  min-height: 100%;
+}
+body {
+  margin-bottom: 60px;
+}
+
+
+/* header */
+
+.header h1 {
+  box-shadow: none;
+  font-family: 'Pacifico', cursive;
+}
+
+.alignCenter {
+  text-align: center;
+}
+
+.pageContent {
+  margin-top: 80px;
+}
+
+/* Landing Page */
+.landing-text {
+  text-align: center;
+}
+
+.raisedButton,
+.raisedButton div,
+.raisedButton button {
+  box-shadow: none;
+}
+
+.raisedButton div {
+  margin: 10;
+  background-color: rgb(210,180,140);
+}
+
+.footer a, .footer a:visited {
+  text-decoration: none;
+  color:  #26C6DA;
+}
+
+.footer a:hover, .footer a:active {
+  text-decoration: underline;
+}
+
+.footer {
+  background-color: rgb(0,96,100);
+  color: #FAFBF6;
+  font-weight: 0;
+//sticky footer
+  position: absolute;
+  bottom: 0;
+  width: 100%;
+  /* Set the fixed height of the footer here */
+  height: 60px;
+  line-height: 60px; /* Vertically center the text there */
+}
+
+/* container for communities */
+.card-container {
+  display: flex;
+}
+
+/* on a single community chain */
+.swipe-btn-up-down {
+  top:50%;
+  left:50%;
+}
+
+.swipe-btn-left-right {
+  margin: 0 auto;
+  vertical-align: middle;
+}
+
