+++ package.json
@@ -46,8 +46,10 @@
     "chai-enzyme": "^0.5.2",
     "chalk": "^2.1.0",
     "check-node-version": "^1.1.2",
+    "classnames": "^2.2.5",
     "concurrently": "^3.1.0",
     "cookie-session": "^2.0.0-alpha.1",
+    "css-loader": "^0.28.7",
     "enzyme": "^2.5.1",
     "express": "^4.14.0",
     "finalhandler": "^1.0.0",
@@ -56,6 +58,7 @@
     "immutable": "^3.8.1",
     "lodash": "^4.17.4",
     "material-ui": "^0.19.2",
+    "node-sass": "^4.5.3",
     "passport": "^0.3.2",
     "passport-facebook": "^2.1.1",
     "passport-github2": "^0.1.10",
@@ -67,15 +70,18 @@
     "react-dom": "^15.3.2",
     "react-quill": "^1.1.0",
     "react-redux": "^5.0.6",
+    "react-router": "^4.2.0",
     "react-router-dom": "^4.2.2",
     "react-swipe": "^5.0.8",
     "redux": "^3.7.2",
     "redux-devtools-extension": "^2.13.0",
     "redux-logger": "^3.0.6",
     "redux-thunk": "^2.2.0",
+    "sass-loader": "^6.0.6",
     "sequelize": "^3.24.6",
     "sinon": "^1.17.6",
-    "sinon-chai": "^2.8.0"
+    "sinon-chai": "^2.8.0",
+    "style-loader": "^0.18.2"
   },
   "devDependencies": {
     "babel": "^6.5.2",
