+++ secrets.js
@@ -1,0 +1,1 @@
+const secret = 'secret'
